+++ src/hexsample/caldb.py
@@ -79,7 +79,7 @@ class CalDB:
         return cls._open(CalibrationType.EQUALIZATION, designator)
 
     @classmethod
-    def open_mle(cls, designator: str) -> PositionCalibrationData:
-        """Open the MLE calibration file for the given designation.
+    def open_position(cls, designator: str) -> PositionCalibrationData:
+        """Open the position calibration file for the given designation.
         """
         return cls._open(CalibrationType.POSITION, designator, PositionCalibrationData)